backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
openai>=1.12.0
aiosqlite>=0.19.0
httpx>=0.26.0
